+++ .gitignore
@@ -10,6 +10,8 @@ db.sqlite3
 # Virtual Environment
 venv/
 env/
+myenv
+
 *.venv
 Pipfile
 Pipfile.lock
@@ -37,3 +39,4 @@ media/
 
 # Windows
 Thumbs.db
+
