+++ api/models/__init__.py
@@ -1,0 +1,4 @@
+from .user import CustomUser
+from .curriculum import Curriculum, CurriculumMapping
+from .course import Course
+from .note import Note
