+++ api/models/user.py
@@ -1,0 +1,20 @@
+from django.db import models
+from django.contrib.auth.models import AbstractUser  # ✅ Import ให้ถูกต้อง
+
+
+
+# ✅ Custom User Model (Extends AbstractUser)
+class CustomUser(AbstractUser):
+    ROLE_CHOICES = (
+        ('admin', 'Admin'),
+        ('member', 'Member'),
+        ('guest', 'Guest'),
+    )
+    role = models.CharField(max_length=10, choices=ROLE_CHOICES, default='guest')
+
+    class Meta:
+        verbose_name = "User"
+        verbose_name_plural = "Users"
+
+    def __str__(self):
+        return f"{self.username} ({self.role})"
