+++ backend/settings.py
@@ -134,4 +134,4 @@ STATIC_URL = 'static/'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 # Custom User Model
-AUTH_USER_MODEL = 'api.CustomUser'
+AUTH_USER_MODEL = 'api.CustomUser'
